test(reporter-solid): move to uvu
