+++ packages/reporter-solid/package.json
@@ -32,36 +32,55 @@
   ],
   "scripts": {
     "prebuild": "rimraf ./dist",
-    "build": "pnpm prebuild && rollup -c",
+    "build": "pnpm prebuild && NODE_ENV=production rollup -c",
     "dev": "rollup -cw",
     "prepublishOnly": "pnpm run build",
-    "test": "jest",
-    "test:ci": "jest --ci --coverage"
+    "test": "node --conditions browser node_modules/uvu/bin.js -r solid-register tests -i common -i setup",
+    "test:ci": "nyc pnpm test"
   },
   "dependencies": {
     "@felte/common": "workspace:*"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
+    "@babel/core": "^7.17.0",
     "@babel/preset-typescript": "^7.14.5",
+    "@babel/register": "^7.17.0",
     "@felte/solid": "workspace:*",
     "@rollup/plugin-babel": "5.3.0",
     "@rollup/plugin-node-resolve": "13.0.5",
-    "@testing-library/jest-dom": "^5.11.9",
     "@types/node": "^15.12.4",
-    "babel-preset-solid": "^1.0.0",
+    "babel-preset-solid": "^1.3.5",
+    "nyc": "^15.1.0",
+    "regenerator-runtime": "^0.13.9",
     "rollup": "^2.52.1",
-    "jest": "^26.6.3",
     "rollup-plugin-terser": "^7.0.2",
     "solid-js": "^1.2.0",
+    "solid-register": "^0.1.5",
     "solid-testing-library": "^0.2.1",
     "tsc-watch": "^4.4.0",
-    "typescript": "~4.3.4"
+    "typescript": "~4.3.4",
+    "uvu": "^0.5.3"
   },
   "peerDependencies": {
     "solid-js": "^1.2.0"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "solid-register": {
+    "compile": {
+      "solid": {
+        "engine": "babel",
+        "extensions": [
+          ".js",
+          ".jsx",
+          ".ts",
+          ".tsx"
+        ]
+      }
+    },
+    "aliases": {
+      "solid": "browser"
+    }
   }
 }
