+++ packages/reporter-solid/tests/setupTests.ts
@@ -1,2 +1,1 @@
 import 'regenerator-runtime/runtime';
-import '@testing-library/jest-dom';
